[training] Make GB200 hero checkpoint writes opt in
Keep benchmark-only model construction in the benchmark and use explicit hero trainer settings. Normal FSDP hero configs skip checkpoint writes unless save_checkpoints is set; the throughput and benchmark launches opt in. One-process JAX clients no longer receive a heartbeat timeout because they have no peer liveness to detect.
