+++ experiments/grug/moe_hero_fsdp/heuristic.py
@@ -109,44 +109,3 @@ def build_hero_configs(*, num_train_steps: int, batch_size: int) -> tuple[GrugMo
         seq_len=model.max_seq_len,
     )
     return model, optimizer
-
-
-def build_checkpoint_benchmark_configs(
-    *, num_train_steps: int, batch_size: int
-) -> tuple[GrugModelConfig, GrugMoeMuonHConfig]:
-    """A 52.85B-total, approximately 1.71B-active MoE for checkpoint measurements."""
-    hidden_dim = 2048
-    model = GrugModelConfig(
-        vocab_size=128_256,
-        hidden_dim=hidden_dim,
-        intermediate_dim=hidden_dim,
-        shared_expert_intermediate_dim=hidden_dim,
-        num_shared_experts=1,
-        num_experts=128,
-        num_experts_per_token=1,
-        num_layers=32,
-        num_heads=16,
-        num_kv_heads=4,
-        local_kv_heads=4,
-        global_kv_heads=2,
-        head_dim=128,
-        max_seq_len=2048,
-        sliding_window=512,
-        global_every=4,
-        capacity_factor=1.0,
-        initializer_std=0.5 / math.sqrt(hidden_dim),
-        qk_mult=1.3,
-        sconv=True,
-        attention_implementation="gpu_fa4_cute",
-        moe_implementation="sonic_cute",
-        expert_chunks=4,
-        report_capacity_overflow=True,
-        rope_fused=True,
-    )
-    optimizer = MoeHeuristic().build_optimizer_config(
-        num_train_steps=num_train_steps,
-        batch_size=batch_size,
-        hidden_dim=model.hidden_dim,
-        seq_len=model.max_seq_len,
-    )
-    return model, optimizer
